# PY3 optimize ChunkedTransferMiddleware and port it to Python 3