- port ChunkedTransferMiddleware to Python 3;
- fixed O(N^2) body building in decode_chunked_transfer;
- fixed O(N^2) string splitting in decode_chunked_transfer.

I also noticed that `decode_chunked_transfer` is not standard compliant. Twisted version should be compliant, but it has the same O(N^2) worst case complexity, and it uses a private `_ChunkedTransferDecoder`:

``` py
from twisted.web.http import _ChunkedTransferDecoder
def decode_chunked_transfer_twisted(chunked_body):
    parts = []
    dec = _ChunkedTransferDecoder(parts.append, lambda _:None)
    dec.dataReceived(chunked_body)
    return b''.join(parts)
```

Benchmark (using an adapted example from our test case):

``` py
chunked_body = b"25\r\n" + b"This is the data in the first chunk\r\n\r\n"
chunked_body += b"1C\r\n" + b"and this is the second one\r\n\r\n"
chunked_body += b"3\r\n" + b"con\r\n"
chunked_body += b"8\r\n" + b"sequence\r\n"

chunked_body = chunked_body * 10000

chunked_body += b"0\r\n\r\n"
```

`chunked_body` size is about 1MB; 
- new version: 48.1 ms
- old version: 2.81 s
- Twisted version: 4.48 s

for 2MB body:
- new version: 111 ms
- old version: 12.3 s
- Twisted version: 21.3 s

It looks like Scrapy http11 handler handles chunked transfer encoding itself (via Twisted) and this change has no effect on it. 

Also, Twisted passes body in chunks. So Twisted version can be modified - this one works much better, it parses 2MB body in 418 ms (not in 21s):

``` py
def decode_chunked_transfer_twisted(chunked_body):
    parts = []
    dec = _ChunkedTransferDecoder(parts.append, lambda _:None)
    size = 10000
    pos = 0
    while pos < len(chunked_body):
        dec.dataReceived(chunked_body[pos:pos+size])
        pos += size
    return b''.join(parts)
```