@redapple 2016-03-31
nice improvement @kmike !
worth contributing back to Twisted?