@kmike 2016-03-31
@redapple Twisted provides `_ChunkedTransferDecoder`; `decode_chunked_transfer_twisted` is not a Twisted function, it is an example on how to implement our `decode_chunked_transfer` using `_ChunkedTransferDecoder`. As last example shows, it can be pretty fast if body is passed in parts, and I think  (not 100% sure) Twisted does that already. It should be possible to rewrite `_ChunkedTransferDecoder` to not use `text.split` (a main cause of O(N^2) complexity) though.