@kmike 2016-09-09
For the record: this PR stalled because it looks like ChunkedTransferMiddleware is not used by Scrapy at all, and we should just remove it.