@redapple 2016-12-08
@kmike , I believe we can close this now that #2411 is merged?